pyproject.toml
[tool.poetry]
name = "hexlet-code"
version = "0.1.0"
description = ""
authors = ["Dmitri Spirin <yamsky.xbox@gmail.com>"]
packages = [
    {include = "gendiff"}
]

[tool.poetry.dependencies]
python = "^3.8"
pytest = "^7.1.2"
coverage = "^6.4.1"
pytest-cov = "^3.0.0"
PyYAML = "^6.0"

[tool.poetry.dev-dependencies]
flake8 = "^4.0.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
gendiff = "gendiff.scripts.gendiff:main"